+++ assets/vue/api.js
@@ -7,7 +7,8 @@ import {
     SubscribersClient,
     SubscriptionClient,
     SubscriberAttributesClient,
-    TemplatesClient
+    TemplatesClient,
+    BouncesClient,
 } from '@tatevikgr/rest-api-client';
 
 const appElement = document.getElementById('vue-app');
@@ -32,6 +33,7 @@ export const statisticsClient = new StatisticsClient(client);
 export const subscriptionClient = new SubscriptionClient(client);
 export const subscriberAttributesClient = new SubscriberAttributesClient(client);
 export const templateClient = new TemplatesClient(client);
+export const bouncesClient = new BouncesClient(client);
 
 export const fetchAllLists = async ({ limit = 100, maxPages = 100 } = {}) => {
     const lists = [];
@@ -53,35 +55,6 @@ export const fetchAllLists = async ({ limit = 100, maxPages = 100 } = {}) => {
     }
 
     return lists;
-};
-
-export const fetchAllBounces = async ({ limit = 100, maxPages = 100 } = {}) => {
-    const bounces = [];
-    let afterId = null;
-
-    for (let pageIndex = 0; pageIndex < maxPages; pageIndex += 1) {
-        const query = { limit };
-        if (afterId !== null) {
-            query.after_id = afterId;
-        }
-
-        const response = await client.get('bounces', query);
-        const items = Array.isArray(response?.items) ? response.items : [];
-        bounces.push(...items);
-
-        const pagination = response?.pagination ?? {};
-        const hasMore = pagination.hasMore === true || pagination.has_more === true;
-        const nextCursor = pagination.nextCursor ?? pagination.next_cursor ?? null;
-
-        if (!hasMore || !Number.isFinite(nextCursor) || nextCursor === afterId) {
-            break;
-        }
-
-        afterId = nextCursor;
-    }
-
-    bounces.sort((a, b) => Number(b.id ?? 0) - Number(a.id ?? 0));
-    return bounces;
 };
 
 export default client;
